refactor: MapPage.jsx 리팩토링 및 테스트페이지 mock 데이터 분리(#21)
