+++ src/pages/MapPage/MapPage.jsx
@@ -13,35 +13,38 @@ import SearchInput from '../../components/SearchInput/SearchInput'
 import TransportSelector from '../../components/Transport/TransportSelector'
 import TransitTurnByTurnList from '../../components/NavigationGuidance/TransitTurnByTurnList'
 import TurnByTurnList from '../../components/NavigationGuidance/TurnByTurnList'
-import { buildTransitDetailLegs } from '../../components/NavigationGuidance/transitDetailMapper'
-import useNavigationRoute from '../../hooks/useNavigationRoute'
 import { mockMapPois } from '../../mocks/map/poi.mock'
 import { mockSearchPlaces } from '../../mocks/search/searchPage.mock'
 import isRegisteredPlace from '../../utils/map/isRegisteredPlace'
-import {
-  DEFAULT_ROUTE_ORIGIN,
-  toNavigationPlace,
-  toNavigationRequestInput,
-} from '../../utils/map/navigationPlaceMapper'
 import { resolvePoiFromSearch } from '../../utils/map/searchPoiResolver'
+import useMapNavigationController from './useMapNavigationController'
 import './MapPage.css'
 
 export default function MapPage() {
   const location = useLocation()
   const navigate = useNavigate()
   const selectedSearchPlace = location.state?.selectedSearchPlace
-  const [currentNav, setCurrentNav] = useState('map')
-  const [transportMode, setTransportMode] = useState('walk')
-  const [routeSheetOpen, setRouteSheetOpen] = useState(false)
-  const [guidanceStarted, setGuidanceStarted] = useState(false)
-  const [guidanceView, setGuidanceView] = useState('map')
-  const [activeGuidanceStepIndex, setActiveGuidanceStepIndex] = useState(0)
-  const [routeOrigin, setRouteOrigin] = useState(DEFAULT_ROUTE_ORIGIN)
-  const [routeDestination, setRouteDestination] = useState(null)
   const [selectedPoi, setSelectedPoi] = useState(() =>
     resolvePoiFromSearch(selectedSearchPlace, mockMapPois),
   )
-  const navigationRoute = useNavigationRoute()
+  const navigation = useMapNavigationController()
+  const {
+    activeGuidanceStep,
+    boundedGuidanceStepIndex,
+    currentNav,
+    guidanceStarted,
+    guidanceSteps,
+    guidanceView,
+    isIndoorGuidanceStep,
+    isRouteMode,
+    isTransitGuidance,
+    navigationRoute,
+    routeDestination,
+    routeOrigin,
+    routeSheetOpen,
+    transitDetailLegs,
+    transportMode,
+  } = navigation
   const registeredPlaces = useMemo(
     () => mockSearchPlaces.filter((place) => place.isRegistered),
     [],
@@ -50,145 +53,13 @@ export default function MapPage() {
     () => isRegisteredPlace(selectedPoi, registeredPlaces),
     [registeredPlaces, selectedPoi],
   )
-  const isRouteMode = currentNav === 'navigation' || guidanceStarted
-  const guidanceSteps = useMemo(
-    () =>
-      navigationRoute.turnByTurnSteps.length > 0
-        ? navigationRoute.turnByTurnSteps
-        : navigationRoute.activeFloorSteps,
-    [navigationRoute.activeFloorSteps, navigationRoute.turnByTurnSteps],
-  )
-  const boundedGuidanceStepIndex = Math.min(
-    activeGuidanceStepIndex,
-    Math.max(guidanceSteps.length - 1, 0),
-  )
-  const activeGuidanceStep =
-    guidanceSteps[boundedGuidanceStepIndex] ?? guidanceSteps[0] ?? null
-  const isIndoorGuidanceStep =
-    !!activeGuidanceStep?.floorId ||
-    activeGuidanceStep?.type === 'indoor' ||
-    activeGuidanceStep?.mode === 'INDOOR'
-  const isTransitGuidance =
-    navigationRoute.selectedRouteOption?.mode === 'transit' ||
-    navigationRoute.selectedRouteOption?.routeType === 'TRANSIT'
-  const transitDetailLegs = useMemo(
-    () => buildTransitDetailLegs(navigationRoute.selectedRouteOption),
-    [navigationRoute.selectedRouteOption],
-  )
 
   const closePoiSheet = () => {
     setSelectedPoi(null)
   }
 
-  const requestRoute = async ({
-    origin = routeOrigin,
-    destination = routeDestination,
-    mode = transportMode,
-  } = {}) => {
-    if (!origin || !destination) {
-      return
-    }
-
-    setCurrentNav('navigation')
-    setGuidanceStarted(false)
-
-    try {
-      const routeData = await navigationRoute.requestRoute(
-        toNavigationRequestInput({
-          origin,
-          destination,
-          transportMode: mode,
-          includeIndoor: true,
-        }),
-      )
-
-      if (routeData) {
-        setRouteSheetOpen(true)
-      }
-    } catch {
-      setRouteSheetOpen(false)
-    }
-  }
-
-  const handleSelectDeparture = (place) => {
-    const origin = toNavigationPlace(place, 'origin')
-    if (!origin) {
-      return
-    }
-
-    setRouteOrigin(origin)
-    setCurrentNav('navigation')
-
-    if (routeDestination) {
-      requestRoute({ origin, destination: routeDestination })
-    }
-  }
-
-  const handleSelectArrival = (place) => {
-    const destination = toNavigationPlace(place, 'destination')
-    if (!destination) {
-      return
-    }
-
-    setRouteDestination(destination)
-    requestRoute({ destination })
-  }
-
-  const handleSwapRoute = () => {
-    if (!routeDestination) {
-      return
-    }
-
-    const nextOrigin = routeDestination
-    const nextDestination = routeOrigin
-    setRouteOrigin(nextOrigin)
-    setRouteDestination(nextDestination)
-    requestRoute({ origin: nextOrigin, destination: nextDestination })
-  }
-
-  const handleSelectTransportMode = (nextMode) => {
-    setTransportMode(nextMode)
-
-    if (routeDestination) {
-      requestRoute({ mode: nextMode })
-    }
-  }
-
   const openSearchPage = () => {
     navigate('/search')
-  }
-
-  const resetRouteView = () => {
-    setGuidanceStarted(false)
-    setGuidanceView('map')
-    setActiveGuidanceStepIndex(0)
-    setRouteSheetOpen(false)
-    navigationRoute.resetRoute()
-    setCurrentNav('map')
-  }
-
-  const selectGuidanceStep = (step, index) => {
-    setActiveGuidanceStepIndex(index)
-    setGuidanceView('map')
-
-    if (step?.floorId) {
-      navigationRoute.selectFloorplan(step.floorId)
-    }
-  }
-
-  const moveGuidanceStep = (direction) => {
-    const nextIndex = boundedGuidanceStepIndex + direction
-
-    if (nextIndex < 0 || nextIndex >= guidanceSteps.length) {
-      return
-    }
-
-    const nextStep = guidanceSteps[nextIndex]
-    setActiveGuidanceStepIndex(nextIndex)
-
-    if (nextStep?.floorId) {
-      navigationRoute.selectFloorplan(nextStep.floorId)
-    }
   }
 
   const handleSearchInputKeyDown = (event) => {
@@ -219,11 +90,11 @@ export default function MapPage() {
             step={activeGuidanceStep}
             activeIndex={boundedGuidanceStepIndex}
             total={guidanceSteps.length}
-            onBack={() => setGuidanceView('list')}
-            onClose={resetRouteView}
-            onRouteClick={() => setGuidanceView('list')}
-            onPrevious={() => moveGuidanceStep(-1)}
-            onNext={() => moveGuidanceStep(1)}
+            onBack={navigation.openGuidanceList}
+            onClose={navigation.resetRouteView}
+            onRouteClick={navigation.openGuidanceList}
+            onPrevious={() => navigation.moveGuidanceStep(-1)}
+            onNext={() => navigation.moveGuidanceStep(1)}
           />
 
           {isIndoorGuidanceStep && navigationRoute.mapFloors.length > 0 ? (
@@ -232,7 +103,7 @@ export default function MapPage() {
                 buildingName={navigationRoute.selectedFloorplan?.mapType ?? '도면'}
                 floors={navigationRoute.mapFloors}
                 activeFloor={navigationRoute.selectedFloorplan}
-                onSelect={navigationRoute.selectFloorplan}
+                onSelect={navigation.selectFloorplan}
               />
             </div>
           ) : null}
@@ -244,8 +115,8 @@ export default function MapPage() {
                 destination={routeDestination?.name ?? '도착지'}
                 route={navigationRoute.selectedRouteOption}
                 legs={transitDetailLegs}
-                onBack={() => setGuidanceView('map')}
-                onClose={resetRouteView}
+                onBack={navigation.showGuidanceMap}
+                onClose={navigation.resetRouteView}
               />
             ) : (
               <TurnByTurnList
@@ -254,9 +125,9 @@ export default function MapPage() {
                 route={navigationRoute.selectedRouteOption}
                 steps={guidanceSteps}
                 activeStepId={activeGuidanceStep?.id}
-                onBack={() => setGuidanceView('map')}
-                onClose={resetRouteView}
-                onSelectStep={selectGuidanceStep}
+                onBack={navigation.showGuidanceMap}
+                onClose={navigation.resetRouteView}
+                onSelectStep={navigation.selectGuidanceStep}
               />
             )
           ) : null}
@@ -267,15 +138,15 @@ export default function MapPage() {
             <DirectionSearch
               origin={routeOrigin.name}
               destination={routeDestination?.name ?? '도착지'}
-              onSwap={handleSwapRoute}
-              onBack={resetRouteView}
+              onSwap={navigation.swapRoute}
+              onBack={navigation.resetRouteView}
             />
           </div>
 
           <div className="map-page__transport">
             <TransportSelector
               activeMode={transportMode}
-              onSelect={handleSelectTransportMode}
+              onSelect={navigation.selectTransportMode}
             />
           </div>
 
@@ -285,7 +156,7 @@ export default function MapPage() {
                 buildingName={navigationRoute.selectedFloorplan?.mapType ?? '도면'}
                 floors={navigationRoute.mapFloors}
                 activeFloor={navigationRoute.selectedFloorplan}
-                onSelect={navigationRoute.selectFloorplan}
+                onSelect={navigation.selectFloorplan}
               />
             </div>
           ) : null}
@@ -316,7 +187,7 @@ export default function MapPage() {
       ) : null}
 
       {!guidanceStarted ? (
-        <BottomNav currentKey={currentNav} onChange={setCurrentNav} />
+        <BottomNav currentKey={currentNav} onChange={navigation.setCurrentNav} />
       ) : null}
 
       <MapPoiSheet
@@ -325,23 +196,16 @@ export default function MapPage() {
         place={selectedPoi}
         isRegistered={isSelectedPoiRegistered}
         onClose={closePoiSheet}
-        onDeparture={handleSelectDeparture}
-        onArrival={handleSelectArrival}
+        onDeparture={navigation.selectDeparture}
+        onArrival={navigation.selectArrival}
       />
 
-      <BottomSheetBase isOpen={routeSheetOpen} onClose={() => setRouteSheetOpen(false)}>
+      <BottomSheetBase isOpen={routeSheetOpen} onClose={navigation.closeRouteSheet}>
         <BottomSheetRouteOptions
           mode={transportMode}
           options={navigationRoute.routeOptions}
-          onSelectOption={navigationRoute.selectRouteOption}
-          onStartNavigation={(option) => {
-            navigationRoute.selectRouteOption(option)
-            setGuidanceStarted(true)
-            setGuidanceView('list')
-            setActiveGuidanceStepIndex(0)
-            setCurrentNav('navigation')
-            setRouteSheetOpen(false)
-          }}
+          onSelectOption={navigation.selectRouteOption}
+          onStartNavigation={navigation.startNavigation}
         />
       </BottomSheetBase>
     </main>
